+++ packages/platform-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/platform-core
 
+## 1.2.2
+
+### Patch Changes
+
+- 249c0e5: Handle insufficient disk space errors
+
+## 1.2.1
+
+### Patch Changes
+
+- 71ef398: Report npm user agent
+- Updated dependencies [f1db886]
+  - @aws-amplify/plugin-types@1.5.0
+
 ## 1.2.0
 
 ### Minor Changes